+++ assets/js/pairings.js
@@ -270,8 +270,7 @@ async function getPDFText(file, fileName) {
     // The workerSrc property shall be specified.
     //
     pdfjsLib.GlobalWorkerOptions.workerSrc =
-        'https://cdnjs.cloudflare.com/ajax/libs/pdf.js/3.4.120/pdf.worker.min.js';
-    // 'https://cdnjs.cloudflare.com/ajax/libs/pdf.js/2.6.347/pdf.worker.min.js';
+        'https://cdnjs.cloudflare.com/ajax/libs/pdf.js/3.11.174/pdf.worker.min.js';
 
     //
     // Asynchronous download PDF